Refactor code and fix formatting issues
