+++ src/App.js
@@ -27,7 +27,7 @@ function App() {
         <Route path="/reservations" element={<Reservations />} />
         {/* </Route> */}
       </Routes>
-      </Router>
+    </Router>
   );
 }
