+++ src/Components/auth/SignInForm.js
@@ -26,13 +26,11 @@ function Login() {
       const data = await response.json();
       if (response.ok) {
         console.log('Login succes:', data);
-
       } else {
         console.error('Login Error:', data);
 
-        navigate('/games'); 
-      } 
-      
+        navigate('/games');
+      }
     } catch (error) {
       console.error('Login Error:', error);
     }
